Feat: 조도 단일 조회 구현 완료

+++ src/main/java/com/nhnacademy/sensordata/entity/Illumination.java
@@ -22,5 +22,5 @@ public class Illumination {
     @Column(name = "topic")
     private String topic;
     @Column(name = "value")
-    private Double value;
+    private Integer value;
 }
